requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.34
APScheduler==3.10.4
netmiko==4.3.0
paramiko==3.5.0
python-dotenv==1.0.1
requests==2.32.3
puresnmp==1.20.0

